GUACAMOLE-374: Evaluate Jinja template paths consistently.
If Jinja templates include other Jinja templates, those include paths
are interpreted relative to the outermost Jinja invocation. The only way
for this to be consistent is if paths are always interpreted relative to
the top of the source path. Template-relative paths can only work if
templates never import each other.